// src/types/internal.d.ts
export interface ReactPropsUserInfo {
    id_str: string;
    name: string;
    screen_name: string;
    description: string;
    verified: boolean;
    is_blue_verified: boolean;
    protected: boolean;
    created_at: string;

    following: boolean;
    followed_by: boolean;
    blocking: boolean;
    blocked_by: boolean;
    muting: boolean;

    followers_count: number;
    friends_count: number;
    statuses_count: number;
    media_count: number;
    favourites_count: number;
}